Challenge 5.5 submission
Made it so the enemies are pushed away when interacting with the player. The enemies move towards the player's goal. The powerup doesn't last forever. Each wave adds more enemies for the player to go against. Added a boost for the player that can be used every 8 seconds using the Space key.